refactor: replace `ElizaOSAdapter` with `HederaElizaOSToolkit`
- Removed `ElizaOSAdapter` and introduced `HederaElizaOSToolkit` for a more modular and maintainable structure.
- Extracted tool handling logic into the new `HederaAgentKitTool`.
- Updated exports and reorganized files accordingly.
